rm file
